Create synonym model, command line program to test synonyms
Moved the files from scripts/load-songs to scripts/load_songs which is more scala friendly version.

Completed the script to take all the lyrics from the billboard top 100 files, and add them into a single file.

Build a synonym model for the lyric data.

Added a command line program to test the synonyms.